refactor user methods into separate files. Fixed imports
