+++ api/src/models/user/user.schema.ts
@@ -1,15 +1,8 @@
 import { Schema } from "mongoose";
-import { findByGoogleId,
-  findOneOrCreateByGoogleId,
-  registerUser,
-  findByEncryptedEmail,
-  findOneByEncryptedEmail,
-  addConnectionToUser,
-  deleteConnectionFromUser,
-  updateUserProfile,
-  createAndPostThread,
-  getConnectionThreads }
-  from "./user.methods";
+import { findOneOrCreateByGoogleId, findByGoogleId, registerUser, findByEncryptedEmail, findOneByEncryptedEmail } from "./user.auth.methods";
+import { addConnectionToUser, deleteConnectionFromUser } from "./user.connections.methods";
+import { updateUserProfile } from "./user.profile.methods";
+import { createAndPostThread, getConnectionThreads } from "./user.thread.methods";
 
 const UserSchema: Schema = new Schema({
   firstName: String,
